fetch/metadata/window-location-reload.https.sub.html

<!DOCTYPE html>
<html lang="en">
  <meta charset="utf-8">
  <meta name="timeout" content="long">
  <title>HTTP headers on request for navigation via window.location.reload</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.sub.js"></script>
  <body>
  <script>
  'use strict';

  function induceRequest(url) {
    const win = window.open();

    return new Promise((resolve) => {
        addEventListener('message', function(event) {
          if (event.source === win) {
            resolve();
          }
        });

        win.location = url;
      })
      .then(() => win.close());
  }

  const responseParams = {
    mime: 'text/html',
    body: `<script>
      if (location.hash === '#first') {
        opener.postMessage('done', '*');
      } else {
        // First, clear the stash entry that was created by the initial
        // navigation
        fetch(location + '&retrieve')
          .then(() => {
            // Set the hash so following the reload, this same script can
            // recognize that it is executing a second time
            location.hash = 'first';
            // Trigger the navigation under test
            location.reload();
          });
      }
    </${''}script>`
  };

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(key, ['httpsOrigin'], responseParams);

    return induceRequest(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-origin');
        });
  }, 'sec-fetch-site - Same origin');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(key, ['httpsCrossSite'], responseParams);

    return induceRequest(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-origin');
        });
  }, 'sec-fetch-site - Cross-site');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(key, ['httpsSameSite'], responseParams);

    return induceRequest(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-origin');
        });
  }, 'sec-fetch-site - Same site');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(key, ['httpsOrigin', 'httpsCrossSite', 'httpsOrigin'], responseParams);

    return induceRequest(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-origin');
        });
  }, 'sec-fetch-site - Same-Origin -> Cross-Site -> Same-Origin redirect');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(key, ['httpsOrigin', 'httpsSameSite', 'httpsOrigin'], responseParams);

    return induceRequest(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-origin');
        });
  }, 'sec-fetch-site - Same-Origin -> Same-Site -> Same-Origin redirect');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(key, ['httpsCrossSite', 'httpsOrigin'], responseParams);

    return induceRequest(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-origin');
        });
  }, 'sec-fetch-site - Cross-Site -> Same Origin');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(key, ['httpsCrossSite', 'httpsSameSite'], responseParams);

    return induceRequest(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-origin');
        });
  }, 'sec-fetch-site - Cross-Site -> Same-Site');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(key, ['httpsCrossSite', 'httpsCrossSite'], responseParams);

    return induceRequest(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-origin');
        });
  }, 'sec-fetch-site - Cross-Site -> Cross-Site');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(key, ['httpsOrigin', 'httpsOrigin'], responseParams);

    return induceRequest(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-origin');
        });
  }, 'sec-fetch-site - Same-Origin -> Same Origin');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(key, ['httpsOrigin', 'httpsSameSite'], responseParams);

    return induceRequest(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-origin');
        });
  }, 'sec-fetch-site - Same-Origin -> Same-Site');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(key, ['httpsOrigin', 'httpsCrossSite'], responseParams);

    return induceRequest(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-origin');
        });
  }, 'sec-fetch-site - Same-Origin -> Cross-Site');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(key, ['httpsSameSite', 'httpsOrigin'], responseParams);

    return induceRequest(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-origin');
        });
  }, 'sec-fetch-site - Same-Site -> Same Origin');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(key, ['httpsSameSite', 'httpsSameSite'], responseParams);

    return induceRequest(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-origin');
        });
  }, 'sec-fetch-site - Same-Site -> Same-Site');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(key, ['httpsSameSite', 'httpsCrossSite'], responseParams);

    return induceRequest(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-origin');
        });
  }, 'sec-fetch-site - Same-Site -> Cross-Site');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(key, ['httpsOrigin', 'httpOrigin', 'httpsOrigin'], responseParams);

    return induceRequest(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-origin');
        });
  }, 'sec-fetch-site - HTTPS downgrade-upgrade');
  </script>
  </body>
</html>
